rag_project/requirements.txt
openai
ollama
langchain
langchain-openai
langchain-community
langchain-ollama
faiss-cpu
numpy
nltk
unstructured[pdf]
python-dotenv
